add CommentsController
