+++ src/controllers/movie-controller.js
@@ -1,5 +1,6 @@
 import FilmCard from '../components/film-card';
 import DetailsPopup from '../components/details-popup';
+import CommentController from './comment-controller';
 import {render, unrender, getRandomValue} from '../utils';
 import {COMMENT_AUTHORS, MINUS_INDEX, COMMENT_DAY, Position} from '../constants';
 
@@ -12,6 +13,7 @@ class MovieController {
     this._onChangeView = onChangeView;
     this._filmCard = new FilmCard(this._dataFilm);
     this._detailsPopup = new DetailsPopup(this._dataFilm);
+    this._containerComments = this._detailsPopup.getElement().querySelector(`.form-details__bottom-container`);
   }
 
   getState() {
@@ -43,6 +45,11 @@ class MovieController {
     }
   }
 
+  _renderCommentsBlock() {
+    const commentController = new CommentController(this._containerComments, this._dataFilm, this._detailsPopup, this.getState, this._onDataChange);
+    commentController.init();
+  }
+
   init() {
     const onEscKeyDown = (evt) => {
       if (evt.key === `Escape` || evt.key === `Esc`) {
@@ -53,15 +60,6 @@ class MovieController {
         unrender(this._detailsPopup.getElement());
         this._detailsPopup.removeElement();
         document.removeEventListener(`keydown`, onEscKeyDown);
-      }
-    };
-
-    const onAddComment = (evt) => {
-      if (evt.ctrlKey && evt.keyCode === 13) {
-        const formData = this.getFormData();
-        this._dataFilm.comments.push(formData.comment);
-        const data = Object.assign(this._dataFilm, this.getState());
-        this._onDataChange(data, this._dataFilm);
       }
     };
 
@@ -136,16 +134,6 @@ class MovieController {
         document.removeEventListener(`keydown`, onEscKeyDown);
       });
 
-    this._detailsPopup.getElement().querySelector(`.film-details__comment-input`)
-      .addEventListener(`focus`, () => {
-        document.removeEventListener(`keydown`, onEscKeyDown);
-      });
-
-    this._detailsPopup.getElement().querySelector(`.film-details__comment-input`)
-      .addEventListener(`blur`, () => {
-        document.addEventListener(`keydown`, onEscKeyDown);
-      });
-
     this._detailsPopup.getElement().querySelector(`.film-details__control-label--watchlist`)
       .addEventListener(`click`, () => {
         const newState = Object.assign(this.getState(), {isWatchlist: !this._dataFilm.isWatchlist});
@@ -160,6 +148,7 @@ class MovieController {
         const newState = Object.assign(this.getState(), {isViewed});
         const data = Object.assign(this._dataFilm, newState);
         this._onDataChange(data, this._dataFilm);
+        this._detailsPopup.getElement().querySelector(`.form-details__middle-container`).classList.toggle(`visually-hidden`);
       });
 
     this._detailsPopup.getElement().querySelector(`.film-details__control-label--favorite`)
@@ -168,22 +157,6 @@ class MovieController {
         const data = Object.assign(this._dataFilm, newState);
         this._onDataChange(data, this._dataFilm);
       });
-
-    this._detailsPopup.getElement().querySelector(`.film-details__comment-input`)
-      .addEventListener(`focus`, () => {
-        document.addEventListener(`keydown`, onAddComment);
-      });
-
-    this._detailsPopup.getElement().querySelector(`.film-details__comment-input`)
-      .addEventListener(`blur`, () => {
-        document.removeEventListener(`keydown`, onAddComment);
-      });
-
-    this._detailsPopup.getElement().querySelectorAll(`.film-details__emoji-item`).forEach((elem) => {
-      elem.addEventListener(`click`, (evt) => {
-        document.querySelector(`.film-details__add-emoji-label`).innerHTML = `<img src="images/emoji/${evt.target.value}.png" width="55" height="55" alt="emoji">`;
-      });
-    });
 
     this._detailsPopup.getElement().querySelectorAll(`.film-details__user-rating-input`).forEach((elem) => {
       elem.addEventListener(`click`, () => {
@@ -203,10 +176,7 @@ class MovieController {
       });
     }
 
-    if (this._dataFilm.isFilmDetails) {
-      // render(this._popupContainer.getElement(), this._detailsPopup.getElement());
-    }
-
+    this._renderCommentsBlock();
     render(this._container.getElement().querySelector(`.films-list__container`), this._filmCard.getElement());
   }
 }
